fix: display selected rows on table-body component
