+++ components/src/components/button/readme.md
@@ -14,19 +14,6 @@
 | `type`      | `type`      |             | `string`  | `undefined` |
 
 
-## Dependencies
-
-### Used by
-
- - [sdds-table](../table)
-
-### Graph
-```mermaid
-graph TD;
-  sdds-table --> sdds-button
-  style sdds-button fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
