+++ components/src/components/table/table-body-row/readme.md
@@ -16,13 +16,13 @@
 
 ### Used by
 
- - [sdds-table](..)
+ - [sdds-table-body](../table-body)
  - [sdds-table-body-row-expendable](..)
 
 ### Graph
 ```mermaid
 graph TD;
-  sdds-table --> sdds-table-body-row
+  sdds-table-body --> sdds-table-body-row
   sdds-table-body-row-expendable --> sdds-table-body-row
   style sdds-table-body-row fill:#f9f,stroke:#333,stroke-width:4px
 ```
