+++ components/src/components/table/table-body/table-body.scss
@@ -1,3 +1,4 @@
-:host {
+:host,
+sdds-table-body {
   display: table-row-group;
 }
