+++ components/src/components/table/table-footer/readme.md
@@ -12,19 +12,6 @@
 | `footerToTable` | Event to send current pagination back to parent | `CustomEvent<number>` |
 
 
-## Dependencies
-
-### Used by
-
- - [sdds-table](..)
-
-### Graph
-```mermaid
-graph TD;
-  sdds-table --> sdds-table-footer
-  style sdds-table-footer fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
