+++ components/src/components/table/table-header-row/readme.md
@@ -12,19 +12,6 @@
 | `headRowToTable` | Send status of main checkbox in header to the parent, sdds-table component | `CustomEvent<boolean>` |
 
 
-## Dependencies
-
-### Used by
-
- - [sdds-table](..)
-
-### Graph
-```mermaid
-graph TD;
-  sdds-table --> sdds-table-header-row
-  style sdds-table-header-row fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
