+++ components/src/components/table/table-header/readme.md
@@ -5,19 +5,6 @@
 <!-- Auto Generated Below -->
 
 
-## Dependencies
-
-### Used by
-
- - [sdds-table](..)
-
-### Graph
-```mermaid
-graph TD;
-  sdds-table --> sdds-table-header
-  style sdds-table-header fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
