+++ components/src/components/table/table-toolbar/readme.md
@@ -21,19 +21,6 @@
 | `sddsTableSearchTerm` | Used for sending users input to main parent <sdds-table> component | `CustomEvent<any>` |
 
 
-## Dependencies
-
-### Used by
-
- - [sdds-table](..)
-
-### Graph
-```mermaid
-graph TD;
-  sdds-table --> sdds-table-toolbar
-  style sdds-table-toolbar fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
